--- self-magritte/src/ingesters/workouts/node/jefit_api/index.js
@@ -1,11 +1,0 @@
-'use strict';
-
-var request = require('request');
-var cheerio = require('cheerio');
-var fetchSingleDate = require('./lib/fetchSingleDate.js');
-var fetchMostRecent = require('./lib/fetchMostRecent.js');
-
-module.exports = {
-  fetchSingleDate: fetchSingleDate,
-  fetchMostRecent: fetchMostRecent
-};
