--- self-magritte/src/ingesters/workouts/node/jefit_api/test/config.example.json
@@ -1,3 +1,0 @@
-{
-	"userId": ""
-}
