+++ self-magritte/src/ingesters/workouts/node/src/index.js
@@ -1,4 +1,4 @@
-const jefit = require('jefit');
+const jefit = require('./jefit_api');
 const util = require('util');
 
 const args = process.argv.slice(2);
