+++ self-magritte/src/ingesters/workouts/node/src/jefit_api/.npmignore
