+++ self-magritte/src/ingesters/workouts/node/src/jefit_api/test/config.example.json
@@ -1,0 +1,3 @@
+{
+	"userId": ""
+}
