+++ self-magritte/src/ingesters/workouts/node/src/jefit_api/test/test2.js
@@ -1,0 +1,6 @@
+var jefit = require('../index.js');
+var config = require('./config.json');
+
+jefit.fetchMostRecent(config.userId, function (result) {
+  console.log(JSON.stringify(result, null, 2));
+});
